refactor: Move `XmlContainer` conversion to a separate sub-module and dedicated traits.
